+++ frontend/src/app/main/ui/components/editable_label.cljs
@@ -14,8 +14,10 @@
    [rumext.alpha :as mf]))
 
 (mf/defc editable-label
-  [{:keys [value on-change on-cancel editing? disable-dbl-click? class-name]}]
-  (let [input (mf/use-ref nil)
+  [{:keys [value on-change on-cancel editing? disable-dbl-click? class-name] :as props}]
+  (let [display-value (get props :display-value value)
+        tooltip (get props :tooltip)
+        input (mf/use-ref nil)
         state (mf/use-state (:editing false))
         is-editing (:editing @state)
         start-editing (fn []
@@ -53,4 +55,5 @@
                                      :on-blur cancel-editing}]
        [:span.editable-label-close {:on-click cancel-editing} i/close]]
       [:span.editable-label {:class class-name
-                             :on-double-click on-dbl-click} value])))
+                             :title tooltip
+                             :on-double-click on-dbl-click} display-value])))